// web/src/lib/api/generated/models/githubComNgodingvarengMemoriaInternalDeliveryRestDtoCommentResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Book API
 * OpenAPI spec version: 1.0
 */

export interface GithubComNgodingvarengMemoriaInternalDeliveryRestDtoCommentResponse {
  body?: string;
  circle_id?: string;
  created_at?: string;
  id?: string;
  moment_id?: string;
  updated_at?: string;
  user_id?: string;
}